requirements.txt
### 2. `requirements.txt`
streamlit==1.38.0
pandas==2.2.3
ollama==0.3.3
tqdm==4.66.5
pillow==10.4.0
spacy==3.7.6
keybert==0.8.5
sentence-transformers==3.1.1
networkx==3.3
requests==2.32.3  # For any URL fetches